refactor: saved quote only after sending claim
